refactor: find and solve more type stability (#5)
* remove certification/witness_sets/monodromy/overdetermined functionality

* affine_only systems

* clean up

* format

* add modelkit tets back

* CheckConcreteStructs and ExplicitImports

* Fix Aqua improvement

* Fix Aqua and JET

* try fist commonsolve interface

* Fix mixed_volume test by removing show_progress parameter

* Refactor solver interface and update tests

- Updated the solver interface to remove the `iterator_only` parameter and replace it with explicit handling for `ResultIterator`.
- Modified the `solve` function signatures across various algorithms to accommodate the new interface.
- Adjusted the `Result` and `ResultIterator` structures to ensure compatibility with the updated solver interface.
- Enhanced tests in `result_test.jl` and `solve_test.jl` to reflect the changes in the solver interface and ensure proper functionality.
- Removed deprecated code related to the old solver interface.

* try to simplify

* Add Valuation structure for solution path approximation

- Introduced a mutable struct `Valuation` to store and compute the valuation of a solution path `x(t)`.
- Implemented initialization and update functions for the `Valuation` struct.
- Added finite difference methods for derivative computation.
- Included functions to check for finiteness and estimate winding numbers.
- Enhanced logging and display capabilities for the `Valuation` data structure.

* Consolidate repeated system validation into one helper

* Collapse init/solve API boilerplate using shared generic helpers

* AlgorithmOptions to hold shared fields

* format

* remove polyhedral and totaldegree API

* format

* Refactor solve functions to streamline argument handling and improve code readability

* cleanup

* Refactor total degree and polyhedral start system functions for improved clarity and consistency

* Refactor result filtering and iteration logic for improved clarity and functionality

* Refactor polyhedral iterator and problem handling for improved structure and clarity

* fix type stabilities end reported JET failures

* more type instability

+++ src/newton/predictor.jl
@@ -136,8 +136,8 @@ order(predictor::Predictor) = predictor.order
 local_error(predictor::Predictor) = predictor.local_error
 trust_region(predictor::Predictor) = predictor.trust_region
 
-function winding_number!(P::Predictor, m)
-    P.winding_number = m
+function winding_number!(P::Predictor, m::Integer)
+    P.winding_number = Int(m)
     return P
 end
